backend/requirements.txt
django
djangorestframework
crawl4ai
scikit-learn
matplotlib
numpy
pandas
joblib
python-dotenv
groq